+++ app/[locale]/projects/[projectId]/project-funding/__tests__/page.spec.tsx
@@ -8,13 +8,9 @@ import { RichTranslationValues } from 'next-intl';
 expect.extend(toHaveNoViolations);
 
 
-jest.mock('next-intl', () => ({
-  useTranslations: jest.fn(),
-}));
-
 jest.mock('next/navigation', () => ({
   useRouter: jest.fn(),
-  useParams: jest.fn()
+  useParams: jest.fn(),
 }));
 type MockUseTranslations = {
   (key: string, ...args: unknown[]): string;
@@ -49,10 +45,7 @@ describe('ProjectsCreateProjectFunding', () => {
     })
 
     const mockUseParams = useParams as jest.Mock;
-
-    // Mock the return value of useParams
     mockUseParams.mockReturnValue({ projectId: '123' });
-
   });
 
   it('should render the component', async () => {
@@ -76,7 +69,7 @@ describe('ProjectsCreateProjectFunding', () => {
     fireEvent.click(screen.getByLabelText('form.yesLabel'));
     fireEvent.click(screen.getByText('buttons.continue'));
     await waitFor(() => {
-      expect(mockUseRouter().push).toHaveBeenCalledWith('/projects/create-project/funder-search')
+      expect(mockUseRouter().push).toHaveBeenCalledWith('/en-US/projects/123/funding-search');
     })
   });
 
@@ -85,7 +78,7 @@ describe('ProjectsCreateProjectFunding', () => {
     fireEvent.click(screen.getByLabelText('form.noLabel'));
     fireEvent.click(screen.getByText('buttons.continue'));
     await waitFor(() => {
-      expect(mockUseRouter().push).toHaveBeenCalledWith('/projects/123')
+      expect(mockUseRouter().push).toHaveBeenCalledWith('/en-US/projects/123')
     })
   });